+++ src/Frontend/Dfe.Complete/Pages/Projects/AboutTheProject/AboutTheProject.cshtml.cs
@@ -1,0 +1,7 @@
+using Dfe.Complete.Models;
+using MediatR;
+
+namespace Dfe.Complete.Pages.Projects.AboutTheProject
+{
+    public class AboutTheProjectModel(ISender sender) : BaseProjectPageModel(sender);
+}