+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/TaskList.cshtml.cs
@@ -1,0 +1,8 @@
+using Dfe.Complete.Domain.Enums;
+using Dfe.Complete.Models;
+using MediatR;
+
+namespace Dfe.Complete.Pages.Projects.TaskList
+{
+    public class TaskListModel(ISender sender) : BaseProjectPageModel(sender);
+}