+++ src/Tests/Dfe.Complete.Tests/Models/BaseProjectsPageModelTests.cs
@@ -1,6 +1,7 @@
 using Dfe.Complete.Application.Projects.Models;
 using Dfe.Complete.Constants;
 using Dfe.Complete.Domain.Enums;
+using Dfe.Complete.Domain.ValueObjects;
 using Dfe.Complete.Models;
 using Dfe.Complete.Pages.Pagination;
 using Dfe.Complete.Tests.Common.Customizations.Models;
@@ -13,7 +14,7 @@ public class BaseProjectsPageModelTests
 {
     [Theory]
     [CustomAutoData(typeof(ListAllProjectResultModelCustomization))]
-    public void GetProjectSummaryUrl_ShouldReturnConversionTaskListUrl(IFixture fixture)
+    public void GetProjectSummaryUrl_ShouldReturnTaskListUrl(IFixture fixture)
     {
         // Arrange
         var project = fixture.Customize(new ListAllProjectResultModelCustomization
@@ -25,6 +26,22 @@ public class BaseProjectsPageModelTests
 
         // Act
         var result = BaseProjectsPageModel.GetProjectSummaryUrl(project);
+
+        // Assert
+        Assert.Equal(expectedUrl, result);
+    }
+
+    [Theory]
+    [CustomAutoData(typeof(ProjectIdCustomization))]
+    public void GetProjectSummaryUrlById_ShouldReturnTaskListUrl(IFixture fixture)
+    {
+        // Arrange
+        var projectId = fixture.Customize(new ProjectIdCustomization()).Create<ProjectId>();
+
+        string expectedUrl = string.Format(RouteConstants.ProjectTaskList, projectId.Value);
+
+        // Act
+        var result = BaseProjectsPageModel.GetProjectSummaryUrl( projectId);
 
         // Assert
         Assert.Equal(expectedUrl, result);