+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Security in case of vulnerabilities.
 ### Added  
 ### Changed  
 ### Fixed  
+ - Routing for projects merged (`/conversion-project` and `/transfer-project` become `/project`)
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-05-08.217...main) for everything awaiting release